# 3rike-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/drivers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "drivers"
                ],
                "summary": "List all drivers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Driver"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drivers"
                ],
                "summary": "Create driver profile",
                "parameters": [
                    {
                        "description": "Driver",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Driver"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Driver"
                        }
                    }
                }
            }
        },
        "/api/drivers/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "drivers"
                ],
                "summary": "Get driver by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Driver ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Driver"
                        }
                    }
                }
            }
        },
        "/api/investors": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "investors"
                ],
                "summary": "List all investors",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Investor"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "investors"
                ],
                "summary": "Create investor profile",
                "parameters": [
                    {
                        "description": "Investor",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Investor"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Investor"
                        }
                    }
                }
            }
        },
        "/api/investors/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "investors"
                ],
                "summary": "Get investor by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Investor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Investor"
                        }
                    }
                }
            }
        },
        "/api/investors/{id}/investments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "investors"
                ],
                "summary": "List all fractions (investments) owned by an investor",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Investor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Fraction"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/tricycles": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "tricycles"
                ],
                "summary": "List all tricycles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Tricycle"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tricycles"
                ],
                "summary": "Register a new tricycle",
                "parameters": [
                    {
                        "description": "Tricycle",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Tricycle"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Tricycle"
                        }
                    }
                }
            }
        },
        "/api/tricycles/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "tricycles"
                ],
                "summary": "Get tricycle by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tricycle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Tricycle"
                        }
                    }
                }
            }
        },
        "/api/tricycles/{id}/fractionalize": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "tricycles"
                ],
                "summary": "Fractionalize a tokenized tricycle into investor shares",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tricycle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fractions",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "total_fractions": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Tricycle"
                        }
                    }
                }
            }
        },
        "/api/tricycles/{id}/tokenize": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "tricycles"
                ],
                "summary": "Tokenize a tricycle on the Canton ledger",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tricycle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Tricycle"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.Driver": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "credit_score": {
                    "type": "integer"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "weeks_remaining": {
                    "description": "out of 70",
                    "type": "integer"
                }
            }
        },
        "domain.Fraction": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "investor_id": {
                    "type": "integer"
                },
                "price_per_unit": {
                    "type": "number"
                },
                "tricycle_id": {
                    "type": "integer"
                },
                "units": {
                    "type": "integer"
                }
            }
        },
        "domain.Investor": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                },
                "wallet_address": {
                    "type": "string"
                }
            }
        },
...
        "domain.Tricycle": {
            "type": "object",
            "properties": {
                "contract_id": {
                    "description": "Canton ledger contract ID",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "driver_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_ev": {
                    "type": "boolean"
                },
                "make": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "price_usd": {
                    "type": "number"
                },
                "status": {
                    "$ref": "#/definitions/domain.TricycleStatus"
                },
                "total_fractions": {
                    "type": "integer"
                }
            }
        },
        "domain.TricycleStatus": {
            "type": "string",
            "enum": [
                "available",
                "financing",
                "owned",
                "tokenized",
                "fractionalized"
            ],
            "x-enum-varnames": [
                "StatusAvailable",
                "StatusFinancing",
                "StatusOwned",
                "StatusTokenized",
                "StatusFractionalized"
            ]
        },
...
    }
}